rename root name db to productapi
