+++ src/main/java/com/butbetter/applicationservices/productapi/model/ProductBeverageEnum.java
@@ -1,4 +1,4 @@
-package com.butbetter.applicationservices.db.model;
+package com.butbetter.applicationservices.productapi.model;
 
 public enum ProductBeverageEnum {
     beer,
